+++ B9 PWings Fork.sln
@@ -3,14 +3,15 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.31129.286
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "B9_Aerospace_WingStuff", "B9 PWings Fork\B9_Aerospace_WingStuff.csproj", "{0EAEF66F-82B2-4A91-915B-624829CE3868}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "B9_Aerospace_WingStuff", "B9_PWings_Fork\B9_Aerospace_WingStuff.csproj", "{0EAEF66F-82B2-4A91-915B-624829CE3868}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "B9_Aerospace_WingStuff-LGG", "B9 PWings Fork\B9_Aerospace_WingStuff-LGG.csproj", "{A2A67033-5BF1-4857-A7BD-C17425A3BC97}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "B9_Aerospace_WingStuff-LGG", "B9_PWings_Fork\B9_Aerospace_WingStuff-LGG.csproj", "{A2A67033-5BF1-4857-A7BD-C17425A3BC97}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{0611BC7B-D781-443D-86A8-0B74B6009CF2}"
 	ProjectSection(SolutionItems) = preProject
 		B9_Aerospace_ProceduralWings.version = B9_Aerospace_ProceduralWings.version
 		buildRelease.bat = buildRelease.bat
+		Changelog.txt = Changelog.txt
 		deploy.bat = deploy.bat
 		jenkins.txt = jenkins.txt
 		LICENSE = LICENSE
