--- B9 PWings Fork/AssemblyVersion.cs
@@ -1,8 +1,0 @@
-
- // This code was generated by a tool. Any changes made manually will be lost
- // the next time this code is regenerated.
- // 
-  
- using System.Reflection;
-
- [assembly: AssemblyVersion("0.43.0.0")]
