+++ B9_PWings_Fork/StaticWingGlobals.cs
