+++ deploy.bat
@@ -1,5 +1,5 @@
 
-@echo off
+rem @echo off
 
 rem H is the destination game folder
 rem GAMEDIR is the name of the mod folder (usually the mod name)
